+++ lib/application/stamp_rally/state/enter_stamp_rally_result.dart
