+++ lib/presentation/page/stamp_rally/public_stamp_rally_view_page.dart
@@ -5,7 +5,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 import '../../../application/stamp_rally/stamp_rally_service.dart';
 import '../../../application/stamp_rally/state/current_public_stamp_rally.dart';
-import '../../../application/stamp_rally/state/entry_stamp_rally_result.dart';
+import '../../../application/stamp_rally/state/enter_stamp_rally_result.dart';
 import '../../../domain/repository/stamp_rally/entity/stamp_rally.dart';
 import '../../component/async_value_handler.dart';
 import '../../component/delimiter_block.dart';
